src/EMMA.Optimizer/requirements.txt
fastapi
uvicorn
psycopg2-binary
pandas
numpy
confluent-kafka
python-dotenv
